+++ ci4/app/Entities/DeploymentSpecificationInitContainer.php
@@ -10,13 +10,15 @@ use App\Core\Entity;
  * @property int $init_container_id
  * @property InitContainer $init_container
  * @property int $position
+ * @property bool $include_in_migration_job
  */
 class DeploymentSpecificationInitContainer extends Entity {
 
-    public static function Create(int $initContainerId, int $pos): DeploymentSpecificationInitContainer {
+    public static function Create(int $initContainerId, int $pos, bool $includeInMigrationJob): DeploymentSpecificationInitContainer {
         $item = new DeploymentSpecificationInitContainer();
         $item->init_container_id = $initContainerId;
         $item->position = $pos;
+        $item->include_in_migration_job = $includeInMigrationJob;
         $item->save();
         return $item;
     }