+++ ci4/app/Interfaces/DeploymentSpecificationInitContainersRequest.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property DeploymentSpecificationInitContainersRequestItem[] $values
+ */
+interface DeploymentSpecificationInitContainersRequest {
+
+}