+++ ci4/app/Interfaces/DeploymentSpecificationInitContainersRequestItem.php
@@ -1,0 +1,9 @@
+<?php namespace App\Interfaces;
+/**
+ * @property int $initContainerId
+ * @property int $position
+ * @property bool $includeInMigrationJob
+ */
+interface DeploymentSpecificationInitContainersRequestItem {
+
+}