+++ ci4/app/Libraries/DeploymentSteps/CronjobStep.php
@@ -239,9 +239,6 @@ class CronjobStep extends BaseDeploymentStep {
             if (strlen($cronJob->container_image->security_context_run_as_group) > 0) {
                 $container->setAttribute('securityContext.runAsGroup', (int)$cronJob->container_image->security_context_run_as_group);
             }
-            if (strlen($cronJob->container_image->security_context_fs_group) > 0) {
-                $container->setAttribute('securityContext.fsGroup', (int)$cronJob->container_image->security_context_fs_group);
-            }
             $container->setAttribute('securityContext.allowPrivilegeEscalation', (bool)$cronJob->container_image->security_context_allow_privilege_escalation);
             $container->setAttribute('securityContext.readOnlyRootFilesystem', (bool)$cronJob->container_image->security_context_read_only_root_filesystem);
 
@@ -285,6 +282,10 @@ class CronjobStep extends BaseDeploymentStep {
                         'name' => $cronJob->container_image->pull_secret,
                     ],
                 ]);
+            }
+
+            if (strlen($cronJob->container_image->security_context_fs_group) > 0) {
+                $template->setSpec('securityContext.fsGroup', (int)$cronJob->container_image->security_context_fs_group);
             }
 
             $resource = new K8sCronJob();