+++ ci4/app/Libraries/DeploymentSteps/DeploymentStep.php
@@ -357,9 +357,6 @@ class DeploymentStep extends BaseDeploymentStep {
         if (strlen($spec->container_image->security_context_run_as_group) > 0) {
             $container->setAttribute('securityContext.runAsGroup', (int)$spec->container_image->security_context_run_as_group);
         }
-        if (strlen($spec->container_image->security_context_fs_group) > 0) {
-            $container->setAttribute('securityContext.fsGroup', (int)$spec->container_image->security_context_fs_group);
-        }
         $container->setAttribute('securityContext.allowPrivilegeEscalation', (bool)$spec->container_image->security_context_allow_privilege_escalation);
         $container->setAttribute('securityContext.readOnlyRootFilesystem', (bool)$spec->container_image->security_context_read_only_root_filesystem);
 
@@ -423,6 +420,10 @@ class DeploymentStep extends BaseDeploymentStep {
             ->setContainers([
                 $container
             ]);
+
+        if (strlen($spec->container_image->security_context_fs_group) > 0) {
+            $template->setSpec('securityContext.fsGroup', (int)$spec->container_image->security_context_fs_group);
+        }
 
         if (strlen($spec->container_image->pull_secret) > 0) {
             $template->setSpec('imagePullSecrets', [